chore: Sync remote repository v4.92

+++ Sources/StreamChatSwiftUI/ChatChannel/Reactions/MessageActions/DefaultMessageActions.swift
@@ -121,17 +121,15 @@ public extension MessageAction {
             messageActions.append(copyAction)
         }
 
-        if message.isRootOfThread {
-            if isInsideThreadView {
-                let markThreadUnreadAction = markThreadAsUnreadAction(
-                    messageController: messageController,
-                    message: message,
-                    onFinish: onFinish,
-                    onError: onError
-                )
-                messageActions.append(markThreadUnreadAction)
-            }
-        } else if !message.isSentByCurrentUser {
+        if message.isRootOfThread && isInsideThreadView {
+            let markThreadUnreadAction = markThreadAsUnreadAction(
+                messageController: messageController,
+                message: message,
+                onFinish: onFinish,
+                onError: onError
+            )
+            messageActions.append(markThreadUnreadAction)
+        } else if !message.isSentByCurrentUser && channel.canReceiveReadEvents {
             if !message.isPartOfThread || message.showReplyInChannel {
                 let markUnreadAction = markAsUnreadAction(
                     for: message,
@@ -145,8 +143,8 @@ public extension MessageAction {
             }
         }
 
-        if message.isSentByCurrentUser {
-            if message.poll == nil {
+        if message.poll == nil, message.giphyAttachments.isEmpty {
+            if channel.canUpdateAnyMessage || channel.canUpdateOwnMessage && message.isSentByCurrentUser {
                 let editAction = editMessageAction(
                     for: message,
                     channel: channel,
@@ -154,7 +152,9 @@ public extension MessageAction {
                 )
                 messageActions.append(editAction)
             }
-
+        }
+        
+        if channel.canDeleteAnyMessage || channel.canDeleteOwnMessage && message.isSentByCurrentUser {
             let deleteAction = deleteMessageAction(
                 for: message,
                 channel: channel,
@@ -164,7 +164,9 @@ public extension MessageAction {
             )
 
             messageActions.append(deleteAction)
-        } else {
+        }
+
+        if !message.isSentByCurrentUser {
             if channel.canFlagMessage {
                 let flagAction = flagMessageAction(
                     for: message,
@@ -180,7 +182,7 @@ public extension MessageAction {
             if channel.config.mutesEnabled {
                 let author = message.author
                 let currentUser = chatClient.currentUserController().currentUser
-                let isMuted = currentUser?.mutedUsers.contains(message.author) ?? false
+                let isMuted = currentUser?.mutedUsers.contains(where: { $0.id == author.id }) ?? false
                 if isMuted {
                     let unmuteAction = unmuteAction(
                         for: message,
@@ -550,7 +552,7 @@ public extension MessageAction {
         onError: @escaping (Error) -> Void
     ) -> MessageAction {
         let action = {
-            messageController.markThreadUnread() { error in
+            messageController.markThreadUnread { error in
                 if let error {
                     onError(error)
                 } else {
